consolidate threshold check test logic
